ТЗ-11 Add: database. #3

+++ src/main/java/ru/yandex/practicum/filmorate/model/MPA.java
@@ -9,10 +9,10 @@ import lombok.extern.jackson.Jacksonized;
 @Builder
 @Jacksonized
 public class MPA {
-    private Integer id;
+    private Long id;
     private String name;
 
-    public MPA(Integer id, String name) {
+    public MPA(Long id, String name) {
         this.id = id;
         this.name = name;
     }
